refactor: dto 디렉토리 구조 변경
- Attribute, Equipment, Rarity API DTO => searchKey
